feat: :sparkles: Funcionamiento y union del atm con el servidor completado

+++ RPC-RMI/controllers/src/main/java/Bank/IBank.java
@@ -10,12 +10,12 @@ import java.util.List;
 public interface IBank extends Remote {
 
     void accountOpening(
-            long ci, String username, String password, String name, double initialAmount, boolean newUser
+            long ci, String username, String password, String name, float initialAmount, boolean newUser
     ) throws RemoteException, CustomException;
-    List<String> transaction(double amount, String description, long accountNumber, long destAccount) throws RemoteException, CustomException;
-    List<String> withdrawal( double amount, long accountNumber) throws RemoteException, CustomException;
+    List<String> transaction(float amount, String description, long accountNumber, long destAccount) throws RemoteException, CustomException;
+    List<String> withdrawal( float amount, long accountNumber) throws RemoteException, CustomException;
     List<String> deposit(
-            double amount, String description, long accountNumber
+            float amount, String description, long accountNumber
     ) throws RemoteException, CustomException;
     List<TransactionEntity> viewAccount(long accountNumber) throws RemoteException, CustomException;
 }
